backend/requirements.txt
fastapi
uvicorn[standard]

duckdb
pydantic-settings

python-nmap

# optional but recommended
python-dotenv
